+++ src/main/java/org/opensearch/knn/index/engine/KNNLibrary.java
@@ -5,6 +5,7 @@
 
 package org.opensearch.knn.index.engine;
 
+import org.opensearch.Version;
 import org.opensearch.common.ValidationException;
 import org.opensearch.knn.index.SpaceType;
 import org.opensearch.knn.index.remote.RemoteIndexParameters;
@@ -150,7 +151,18 @@ public interface KNNLibrary extends MethodResolver {
         return false;
     }
 
+    /**
+     * Checks if the library is deprecated for a given OpenSearch version.
+     *
+     * @param indexVersionCreated the OpenSearch version where the index is created
+     * @return true if deprecated, false otherwise
+     */
+    default boolean isRestricted(Version indexVersionCreated) {
+        return false; // By default, libraries are not deprecated
+    }
+
     default RemoteIndexParameters createRemoteIndexingParameters(Map<String, Object> indexInfoParameters) {
         throw new UnsupportedOperationException("Remote build service does not support this engine");
+
     }
 }